Timezone corrected. Gonna work on categories.

+++ src/main/java/com/owo/OwoDokan/repository/admin_related/OfferRepository.java
@@ -3,6 +3,6 @@ package com.owo.OwoDokan.repository.admin_related;
 import com.owo.OwoDokan.entity.admin_related.OffersEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OfferRepository extends JpaRepository<OffersEntity, Integer> {
+public interface OfferRepository extends JpaRepository<OffersEntity, Long> {
 
 }
